requirements.txt
asgiref==3.10.0
Django==5.2.7
django-simple-history==3.10.1
numpy==2.3.4
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
